+++ src/eos/radiationProperties/CMakeLists.txt
@@ -1,8 +1,8 @@
 target_sources(ablateLibrary
         PUBLIC
         radiationProperties.hpp
-        Constant.hpp
-        Constant.cpp
-        Zimmer.cpp
-        Zimmer.hpp
+        constant.hpp
+        constant.cpp
+        zimmer.cpp
+        zimmer.hpp
         )
